+++ services/tok/goos/init_windows.go
@@ -19,10 +19,11 @@ Your Drupal development environment is now up and running
 	`)
 
 	console.Println(`🔁  Run "tok watch" in a new terminal window to keep your local and container files in sync`, "-")
-	console.Println(`💻  Run "ssh `+conf.GetConfig().Tokaido.Project.Name+`.tok" to access the Drush container`, "-")
+	console.Println(`💻  Run "ssh `+conf.GetConfig().Tokaido.Project.Name+`.tok" to ssh into the environment`, "-")
+	console.Println(`👀  Run "tok exec" to run commands like 'tok exec drush status'`, "-")
 	console.Println(`🌎  Run "tok open" to open the environment in your browser`, "-")
-	console.Println(`🚀  Run "tok status" check the status of your environment`, "-")
+	console.Println(`🚀  Run "tok status" to check the status of your environment`, "-")
 	fmt.Println(`
-Check out https://docs.tokaido.io/environments for tips on managing your Tokaido environment
+Check out https://tokaido.io/docs/using/ for tips to help you get the most out of your Tokaido environment
 	`)
 }