+++ services/tok/init.go
@@ -31,6 +31,7 @@ func Init() {
 	drupal.CheckSettings()
 	docker.FindOrCreateTokCompose()
 	ssh.GenerateKeys()
+	docker.CreateDatabaseVolume()
 
 	git.IgnoreDefaults()
 
@@ -44,7 +45,7 @@ func Init() {
 
 	if c.System.Syncsvc.Enabled {
 		fmt.Println()
-		console.Println("🔄  Creating a background process to sync your local repo into the Tokaido environment", "")
+		console.Println(`🔄  Creating a background process to sync your local repo into the Tokaido environment`, "")
 
 		unison.CreateSyncService(c.Tokaido.Project.Name, c.Tokaido.Project.Path)
 	}